requirements.txt
crewai
python-dotenv
crewai_tools
groq
exa_py
crewai[tools]
litellm